[04] node auth

+++ contacts/contact.controller.js
@@ -1,4 +1,7 @@
 const Contact = require("./Contact");
+const {
+    Types: { ObjectId },
+} = require("mongoose");
 
 class ContactsController {
     constructor() {}
@@ -24,14 +27,15 @@ class ContactsController {
 
     async _addContact(req, res, next) {
         try {
+            const userId = req.user._id;
             const data = req.body;
-            const existingContact = await Contact.findContactByEmail(data.email);
+            const existingContact = await Contact.findContactByEmail(data.email, userId);
 
             if (existingContact) {
                 return res.status(409).send("Contact with this email already exist");
             }
 
-            await Contact.create(data);
+            await Contact.create({ ...data, owner: userId });
 
             return res.status(201).json(data);
         } catch (err) {
@@ -41,7 +45,16 @@ class ContactsController {
 
     async _getContacts(req, res, next) {
         try {
-            const contacts = await Contact.find();
+            const userId = req.user._id;
+            const contacts = await Contact.paginate(
+                { owner: userId },
+                {
+                    populate: {
+                        path: "owner",
+                        select: "email",
+                    },
+                }
+            );
 
             return res.status(200).json(contacts);
         } catch (err) {
@@ -51,8 +64,12 @@ class ContactsController {
 
     async _getContactById(req, res, next) {
         try {
+            const userId = req.user._id;
             const { id } = req.params;
-            const contact = await Contact.findById(id);
+            const contact = await Contact.findOne({ _id: id, owner: userId }).populate({
+                path: "owner",
+                select: "email",
+            });
 
             if (!contact) {
                 return res.status(404).send("contact was not found");
@@ -66,9 +83,14 @@ class ContactsController {
 
     async _updateContactById(req, res, next) {
         try {
+            const userId = req.user._id;
             const { id } = req.params;
 
-            const updatedContact = await Contact.findContactByIdAndUpdate(id, req.body);
+            const updatedContact = await Contact.findContactByIdAndUpdate(
+                id,
+                req.body,
+                userId
+            );
             if (!updatedContact) {
                 return res.status(404).send("contact was not found");
             }
@@ -81,8 +103,9 @@ class ContactsController {
 
     async _deleteContactById(req, res, next) {
         try {
+            const userId = req.user._id;
             const { id } = req.params;
-            const { deletedCount } = await Contact.deleteOne({ _id: id });
+            const { deletedCount } = await Contact.deleteOne({ _id: id, userId });
 
             if (!deletedCount) {
                 return res.status(404).send("contact was not found");
@@ -92,6 +115,16 @@ class ContactsController {
         } catch (err) {
             next(err);
         }
+    }
+
+    validateId(req, res, next) {
+        const { id } = req.params;
+
+        if (!ObjectId.isValid(id)) {
+            return res.status(400).send("Id is not valid");
+        }
+
+        next();
     }
 }
